Enemy Spawning Added
I made an enemy spawning script. Every 0.5 (the roll rate) seconds, the enemy spawner will try to randomly spawn an enemy. Depending on the value of the Spawn Rate, an enemy will either spawn or not spawn.

Oh, and the spawn rate for the enemies is stored in a scriptable object called 'Spawn Rate'